add script for benchmarking hotspot2d

+++ opencl/hotspot/benchmark.sh
@@ -1,13 +1,19 @@
+
 #!/bin/bash
+if [ "$#" -ne 1 ]; then
+				echo "Illegal number of parameters (enter input size)"
+				exit -1
+fi
 
 # prints kernel runtime in nanoseconds to hotspot.out using kernel event times
-
 rm hotspot.raw
 rm hotspot.out
 
+echo temp_$1
+
 for i in {1..10}
 do
-./hotspot 512 1 1 ../../data/hotspot/temp_512 ../../data/hotspot/power_512 output.out >> hotspot.raw
+./hotspot $1 1 1 ../../data/hotspot/temp_$1 ../../data/hotspot/power_$1 output.out >> hotspot.raw
 done
 
 cat hotspot.raw | grep DEBUG | awk '{print $4}' > hotspot.out
